Convert class level filtering test to new JUnit multiversion test
