+++ subprojects/testing-jvm/src/integTest/groovy/org/gradle/testing/junit/AbstractJUnit4ClassLevelFilteringIntegrationTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 the original author or authors.
+ * Copyright 2023 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,28 +14,11 @@
  * limitations under the License.
  */
 
-
 package org.gradle.testing.junit
 
 import org.gradle.integtests.fixtures.DefaultTestExecutionResult
-import org.gradle.integtests.fixtures.TargetCoverage
-import org.gradle.testing.fixture.JUnitMultiVersionIntegrationSpec
 
-import static org.gradle.testing.fixture.JUnitCoverage.JUNIT_VINTAGE
-import static org.gradle.testing.fixture.JUnitCoverage.LARGE_COVERAGE
-
-@TargetCoverage({ LARGE_COVERAGE + JUNIT_VINTAGE})
-class JUnitClassLevelFilteringIntegrationTest extends JUnitMultiVersionIntegrationSpec {
-
-    def setup() {
-        buildFile << """
-            apply plugin: 'java'
-            ${mavenCentralRepository()}
-            dependencies { ${getDependencyBlockContents()} }
-            test { use${testFramework}() }
-        """
-    }
-
+abstract class AbstractJUnit4ClassLevelFilteringIntegrationTest extends AbstractJUnitClassLevelFilteringIntegrationTest {
     def "runs all tests for class instead of method when runner is not filterable"() {
         file("src/test/java/FooTest.java") << """
             import org.junit.*;
@@ -43,7 +26,7 @@ class JUnitClassLevelFilteringIntegrationTest extends JUnitMultiVersionIntegrati
             @RunWith(SomeRunner.class)
             public class FooTest {
             }
-        """
+        """.stripIndent()
         file("src/test/java/SomeRunner.java") << """
             import org.junit.*;
             import org.junit.runner.*;
@@ -69,7 +52,7 @@ class JUnitClassLevelFilteringIntegrationTest extends JUnitMultiVersionIntegrati
                     }
                 }
             }
-        """
+        """.stripIndent()
 
         when:
         run("test", "--tests", "FooTest.pass")
@@ -90,37 +73,5 @@ class JUnitClassLevelFilteringIntegrationTest extends JUnitMultiVersionIntegrati
 
         then:
         failure.assertHasCause("No tests found for given includes: [NotFooTest.pass]")
-    }
-
-    def "can filter tests from a superclass"() {
-        given:
-        file('src/test/java/SuperClass.java') << '''
-            import org.junit.Test;
-
-            public abstract class SuperClass {
-                @Test
-                public void superTest() {
-                }
-            }
-        '''
-        file('src/test/java/SubClass.java') << '''
-            import org.junit.Test;
-
-            public class SubClass extends SuperClass {
-                @Test
-                public void subTest() {
-                }
-            }
-        '''
-
-        when:
-        succeeds('test', '--tests', 'SubClass.superTest')
-
-        then:
-        new DefaultTestExecutionResult(testDirectory)
-            .assertTestClassesExecuted('SubClass')
-            .testClass('SubClass')
-            .assertTestCount(1, 0, 0)
-            .assertTestPassed('superTest')
     }
 }
