+++ subprojects/testing-jvm/src/integTest/groovy/org/gradle/testing/junit/JUnit4ClassLevelFilteringIntegrationTest.groovy
@@ -1,0 +1,26 @@
+/*
+ * Copyright 2023 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.testing.junit
+
+import org.gradle.integtests.fixtures.TargetCoverage
+import org.gradle.testing.junit4.JUnit4MultiVersionTest
+
+import static org.gradle.testing.fixture.JUnitCoverage.LARGE_COVERAGE
+
+@TargetCoverage({ LARGE_COVERAGE })
+class JUnit4ClassLevelFilteringIntegrationTest extends AbstractJUnit4ClassLevelFilteringIntegrationTest implements JUnit4MultiVersionTest {
+}
